+++ kodilla-patterns2/src/main/java/com/kodilla/patterns2/decorator/pizzaportal/PizzaOrder.java
@@ -1,8 +1,6 @@
 package com.kodilla.patterns2.decorator.pizzaportal;
 
-import java.math.BigDecimal;
-
 public interface PizzaOrder {
-    BigDecimal getCost();
+    double getCost();
     String getDescription();
 }
